[refactor] #133 Remind Live, Test 분리
